+++ es-app/src/views/gamelist/BasicGameListView.h
@@ -36,6 +36,7 @@ protected:
     virtual std::string getQuickSystemSelectLeftButton() override;
     virtual void populateList(const std::vector<FileData*>& files) override;
     virtual void remove(FileData* game, bool deleteFile) override;
+    virtual void removeMedia(FileData* game) override;
     virtual void addPlaceholder();
 
     TextListComponent<FileData*> mList;