+++ es-app/src/views/gamelist/GridGameListView.h
@@ -43,6 +43,7 @@ protected:
     virtual std::string getQuickSystemSelectLeftButton() override;
     virtual void populateList(const std::vector<FileData*>& files) override;
     virtual void remove(FileData* game, bool deleteFile) override;
+    virtual void removeMedia(FileData* game) override;
     virtual void addPlaceholder();
 
     ImageGridComponent<FileData*> mGrid;